+++ src/app/services/xml-parsers/apparatus-entries-parser.service.ts
@@ -1,19 +1,18 @@
 import { Injectable } from '@angular/core';
-import { parse } from '.';
+import { ParserRegister } from '.';
 import { ApparatusEntry, Reading, Witness, XMLElement } from '../../models/evt-models';
-import { AppParser } from './app-parser';
-import { createParser } from './parser-models';
 
 @Injectable({
   providedIn: 'root',
 })
 export class ApparatusEntriesParserService {
   private appEntryTagName = 'app';
-  private appParser = createParser(AppParser, parse);
 
   public parseAppEntries(document: XMLElement) {
+    const appParser = ParserRegister.get('evt-apparatus-entry-parser');
+
     return Array.from(document.querySelectorAll<XMLElement>(this.appEntryTagName))
-      .map((appEntry) => this.appParser.parse(appEntry));
+      .map((appEntry) => appParser.parse(appEntry) as ApparatusEntry);
   }
 
   public getSignificantReadings(apps: ApparatusEntry[]) {
@@ -34,7 +33,7 @@ export class ApparatusEntriesParserService {
     return signRdgsNumber;
   }
 
-  public getAppVariance(signRdgsNum: { [key: string]: number }, witList: { [key: string]: Witness }) {
+  public getAppVariance(signRdgsNum: { [key: string]: number }, witList: Witness[]) {
     const appsVariance = {};
     if (Object.keys(witList).length > 1) {
       Object.keys(signRdgsNum).forEach((x) => {